[SYSTEMML-1791] Performance frame cache release (avoid scans)
This patch removes unnecessary frame cache release overhead for JMLC
deployments. On cache release, we update the size of currently pinned
objects, which is only used in the hybrid spark execution mode. For
frame with string schema this creates unnecessary overhead because for
string column this is not a meta data operation but requires a full scan
over the individual strings to determine their lengths. We now disable
this in forced CP execution mode, which significantly improved
performance in the context of column indexing, which are meanwhile pure
meta data operations.