[SYSTEMML-1796] Provide fine-grained statistics for SystemML
Though SystemML provides explain and statistics tooling functionality to
understand the execution plan and the performance, most users (even
committers) are not able to explain why SystemML's performance is so good
or bad. This is likely because SystemML's optimizer performs sophisticated
dataflow analysis, cost optimization as well as static and dynamic
rewrites. New features/backends such as code generation, automatic fusion,
compression and GPU backend adds additional level of complexity to the
optimizer. This commit is a small step in the direction of explaining
SystemML's behavior.

Instead of examining statistics output and then
trying to correlate with the explain plan, this commit provides
instruction-level statistics with the corresponding DML snippet (file
beginLine:beginCol-endLine-endCol).

Output with the configuration property systemml.stats.finegrained set to true:
```
Heavy hitter instructions:
  #  Instruction               Time(s)  Count  Misc Timers
  1  solve [148:0-148:0]         0.027      1  aqmd[0.000s,1]
  2  tsmm [137:4-137:13]         0.008      1  rlsi[0.000s,1], aqmd[0.000s,1], aqrd[0.000s,1]
  3  uak+ [161:10-161:10]        0.007      1  aqrd[0.000s,1], rlsi[0.000s,1]
  4  ba+* [138:4-138:13]         0.005      1  aqmd[0.000s,1], aqrd[0.000s,2], rlsi[0.000s,2]
  5  rand [89:0-89:0]            0.004      1  aqmd[0.000s,1]
  6  * [133:9-133:24]            0.002      1  aqmd[0.000s,1], aqrd[0.000s,1], rlsi[0.000s,1]
  7  r' [138:4-138:13]           0.001      2  aqmd[0.000s,2], aqrd[0.000s,2], rlsi[0.000s,2]
  8  append [97:4-97:4]          0.001      1  rlswr[0.000s,1], aqmd[0.000s,1], rlsev[0.000s,0], aqrd[0.000s,2], rlsi[0.000s,2]
  9  + [144:4-144:21]            0.001      1  aqmd[0.000s,1], aqrd[0.000s,2], rlsi[0.000s,2]
 10  leftIndex [104:4-104:26]    0.001      1  aqmd[0.000s,1], rlsi[0.000s,1], aqrd[0.000s,1]
```

aqrd and aqmd refers to dense input and outputs respectively. Similarly, aqrs and aqms refers to sparse input and outputs. It can also show both:
```
1  leftIndex [106:4-106:42]                                                       141.928   28552  aqrs[0.024s,23165], rls[0.066s,57104], csrlix[116.652s,22649], s2d[9.297s,258], aqrd[14.097s,33939], lixcp[1.410s,5903]
```

Closes #577.