[SYSTEMML-1818] Fix ctable size propagation w/ two sequence inputs
For ctable, we exploit the knowledge of sequence inputs for better size
propagation. This patch fixes an issue of incorrect size propagation,
where the sequence rewrite applies to the right, but we mistakenly
leverage the size of the left input if it is a sequence as well. An
example for this problematic pattern is (from Kmeans train):

R = table(seq(i, NC*(NR-1)+i, NC), seq(1, NR, 1));

Furthermore, this patch also fixes the unnecessary eager spark context
creation in forced singlenode execution mode and globally renames any
remaining occurrences of tertiary operators to ternary operators.