[SYSTEMML-1833] Fix codegen race conditions and cleanup robustness
This patch fixes two codegen issues which led to failures on the ARIMA
and MDABivariate algorithms. In detail, this includes (1) a fix of race
conditions on variable name creation/usage in multi-threaded parfor
contexts with dynamic recompilation, and (2) a fix of cleanup robustness
for row-wise templates with partial unknowns (which led to selecting a
row template for column vectors and thus invalid generated source code).
Additionally, we now also avoid unnecessary dense allocations for empty
side vector inputs to row-wise templates because these are only used
through cell indexing or dot products. 

Furthermore, this also includes a cleanup of the DML and R scripts for
MDABivariate statistics (vectorized loops), which simplifies debugging
and improves the R performance of these tests by almost 2x.