[SYSTEMML-1834] Codegen plan enumeration with partial costing
This PR adds conditions to the fusion plan enumerator to stop costing a
plan when its partially computed cost exceeds the cost of the best plan
found so far (an upper bound).

The Statistics class reports the number of partially costed plans
separately from the number of fully costed plans. I attempted to count
the number of Hops costed per plan in order to measure the effectiveness
of stopping costing early at a more fine-grained level, but this count
was inaccurate for some reason. On StratStats, 2301 out of 3742 plans
can be stopped costed early.

Closes #611.