[SYSTEMML-1838] Performance sparse/ultra-sparse binary block read
This patch makes a number of performance improvements for CP read
(single- and multi-threaded) of sparse and ultra-sparse matrices in
binary block format:

(1) Allocate reuse block in CSR format: This avoids unnecessary
reallocations in the presence of a mix of sparse and ultra-sparse blocks
because ultra-sparse blocks clear any reuse block that is not in CSR.

(2) Special CSR init for sparse blocks from input streams (in addition
to the existing init for ultra-sparse blocks)

(3) Exploit estimated number of non-zeros on sparse block append (also
used during read), which changes the allocation factors from 1.1x to 2x
until the estimated number of non-zeros per row is reached and hence,
avoids unnecessary reallocations.

(4) Improved load balance via more tasks on the final sort of sparse
rows.

Together these changes improved the CP read time of a 1M x 1M matrix
with sparsity 0.001 (with a mix of sparse and ultra-sparse blocks) of
roughly 12GB in size and 1M blocks from 41s to 32.4s.