[SYSTEMML-1851] Fix transformdecode output column names (from meta)
This patch fixes the output schema handling of spark and cp
transformdecode instructions. For transformdecode the output column
names were always set to default names (e.g., C1, C2, etc) because the
input is a matrix. We now properly initialize the output column names,
based on the column names of the provided meta data frame.