[SYSTEMML-1862] Fix special case buffer pool eviction, cleanup
This patch fixes a very specific special case of buffer pool eviction,
where an intermediate is exactly of the size of the buffer pool. There
was a mismatch of conditions regarding entering the eviction path, and
eviction continuation (<= vs <). In such a scenario, we would try to
evict the next entry although there are no more entries to evict and
hence run into a nosuchelement exception. 

Furthermore, this patch also cleans up the maintenance of "fine-grained
statistics" by moving it out of the lazy write buffer.