[SYSTEMML-1871] Rework compiler/runtime predicate handling (if, loops)
So far we handled literal and expression predicates differently in
blocks such as if, while, for, and parfor. This was because DAGs of data
ops and literals do not compile to any instructions. The two different
code paths unnecessarily complicated the compiler and runtime,
especially with regard to rewrites such as constant folding, which can
cause a transition between the two code paths. We now properly compile a
transient write HOP on top of any predicate DAGs, which unifies two
cases and simplifies both compiler and runtime. 

Furthermore, this patch also includes improvements for determining the
number of iterations during parfor optimization for loops inside the
surrounding parfor body, as well as a couple of minor cleanups.