[SYSTEMML-1888] Fix parfor optimizer exec type selection (msvm,kmeans)
This patch makes a number of smaller fixes to the parfor optimizer
execution type selection to enable the compilation of remote parfor
plans in the presence of many inner iterations but relatively small
input data. In detail, this includes (besides various cleanups): 

(1) Fixes for determining the number of inner iterations (include for
and while loops with defaults if necessary), 

(2) A modified minimum datasize threshold that is now scaled by the
number of inner iterations, and

(3) A better handling of what-if memory estimation (e.g., all operations
to CP) that does not account for unnecessary hops that are never
compiled to lops due to hop-lop rewrites.

On the perftest MSVM 1M x 1K, sparse scenario with 150 classes and 25
iterations, this patch improved performance from 288s to 94s on a 1+6
node cluster (including read and spark context creation).