[SYSTEMML-1894] Performance log operations (replace FastMath by Math)
SystemML uses commons-math's FastMath instead of Math for certain
builtin operations due to performance and/or accuracy reasons. However,
recent experiments with jdk8 on different platforms showed that for log
operations this is meanwhile actually counter-productive. Hence, this
patch globally replaces FastMath.log with Math.log. 

On a scenario of 100 iterations of log(X), where X is a 10Kx10K dense
matrix, this patch improved performance from 413.1s to 319.8s.