[SYSTEMML-1904] Improved codegen dynamic recompilation decisions
This patch improves the decision on marking HOPs and thus, HOP DAGs for
dynamic recompilation. For example, on AutoEncoder, the indexing of
batches causes unknowns in the main DAG of the inner loop. Since the
worst-case estimates are sufficient to compile these operations to CP,
this DAG is not marked for dynamic recompilation. Hence, the codegen
optimizer can never re-optimize this DAG again. We now use more
aggressive recompilation decisions for codegen, effectively marking
every HOP DAG with unknown dimensions for dynamic recompilation. 

On one epoch of AutoEncoder over Mnist60k, batch=512, H1=500, H2=2, this
change improved the end-to-end performance from 39s to 24s.

Furthermore, this also includes two minor improvements of codegen outer
templates: (1) improved performance of full aggregations due to less
array loads and stores, and (2) a fix of the fusion condition to avoid
transpose operations on the main input and intermediates.