[SYSTEMML-1743] Create lightweight SystemML artifact
Create a lightweight SystemML jar file based on the execution of
a particular set of code, such as JMLC. In addition to generating
a minimal jar file that corresponds to this code, automatically
generate maven assembly dependency sets that can be used to build
a 'lite' jar that can be activated via a 'lite' maven profile.
This essentially creates an minimum-size uberjar that can perform
tasks such as scoring with JMLC, where distributed operations on
Spark and Hadoop are not desired.

Closes #578.