[SYSTEMML-1918] Fix rewrite 'merge block sequence' (tmp invalid DAGs)
This patch fixes the statement block rewrite RewriteMergeBlockSequence,
which caused temporarily invalid HOP DAGs (e.g., twrites without inputs)
in the original program because the list of statement blocks is modified
and only finally attached back to the statement block hierarchy. This
creates problems for other statement block rewrites, which need to make
a full pass over the program for global analysis.

Furthermore, this patch also eliminates a debug condition, which
unnecessarily disabled this rewrite for special cases.