[SYSTEMML-1914] Extended compression rewrite (auto compression, default)
This patch extends the existing compression rewrite (true/false,
statically applied to persistent reads of multi-column matrices) by auto
compression. If configured with 'auto' (which is enabled by default), we
now automatically decide to compress or not. Initially, we follow a
conservative approach: we decide to compress matrices that are known to
exceed aggregate cluster memory, that are not ultra sparse, and only if
they are used in loops and all operations are supported over compressed
matrices. In order to analyze the individual operations, we make a full
pass over the program with awareness of functions and transitive
assignments of compressed matrices. 

On L2SVM (10 outer iterations, tol=10^-9) and Mnist240m (~540GB in
uncompressed serialized form), this change improved performance from
5,913s to 1,952s. At the same time, the overhead for program analyze is
negligible taking 9ms and 0.8ms on L2SVM for the first and second
rewrite applicable. This program analysis is conditional on all the
other conditions (incl large out-of-core matrices) and thus, do not
affect the compilation of programs over small data.

Furthermore, this patch also makes a couple of minor cleanups such as
the handling of function keys, variable sets, and hop primitives.