[SYSTEMML-1921] Bug in backward of recurrent layers when only returning final output
This fixes a typo in the `backward` functions of the recurrent layers
that causes an error when `given_sequences` is false.  Specifically,
when only the final time-step's output was desired from the forward
functions, `out` and `dout` will be of shape `(N, M)`, and within
the `backward` functions, `dout` will be padded with zeros for the
earlier time-steps to enlarge the shape to `(N, T*M)`.  While the goal
of the logic and the associated comments were correct, there was a typo
in which the shape was instead enlarged to `(T*D)`.

This also updates the associated gradient checks to also run the same
test with `return_sequences = FALSE`.