[SYSTEMML-1922] Fix memory-efficiency spark broadcasts, OOM AutoEncoder
This patch fixes OOM issues on our staging/autoencoder-2layer.dml
script, which revealed a general issue of unaccounted broadcast memory
requirements. In detail, Spark keeps a deep copy of the blockified
broadcast as well as a reference to the passed object. So far we did not
account for the latter, which keeps this object in-memory even though
its evicted and put onto a soft reference. 

We now explicitly clear blocks (except in local mode) after the
broadcast has been created (and thus blockified) because remote fetches
only reference the blockified chunks.