[SYSTEMML-1932] Fix update-in-place compiler/runtime for frames, build
This patch fixes update-in-place issues with leftindexing operations
over frames. In detail, this includes additional data type checks in the
rewrite for selecting update-in-place candidates, and a robustness
improvement of the respective runtime to simply ignore candidates of
wrong data types.

Furthermore, this also includes an additional fix of several testsuite
configuration files because our jenkins server expects a local
scratch_space and local tmp dir per testcase.