[SYSTEMML-1933] Generalized codegen cbind handling in row-wise ops
This patch generalizes the compilation of cbind operations in codegen
row templates. So far, we only supported cbind with a vector of zeros,
and cbind closed the row template. We now support cbind operations (with
vectors of arbitrary constants) in the middle of row templates, which
also allows for multiple cbind operations in a single fused operator.