[SYSTEMML-1934] Fix codegen optimizer (incorrect structural pruning)
This patch fixes a severe codegen optimizer issue, where in special
cases the positions of sub problems were incorrectly set leading to
wrong mappings of optimal plans for subproblems to the global plan. We
now use a much simpler and more robust creation of these mappings. 

With this patch we now (1) find the optimal plans on scenarios where we
previously missed them (e.g., Mlogreg), and (2) structural pruning shows
more pruning effectiveness. For example, on GLM binomial probit, there
are 264,371 plans - with cost-based pruning this is reduced to 33,388
and with additional structural pruning further reduced to 9,574 plans.

Furthermore, this patch also improves the trace information of the
codegen optimizer as well as fixes an issue of statistic maintenance
without structural pruning.