[SYSTEMML-1904] Improved codegen for recompile-once functions
This patch improves the reset of recompilation flags on dynamic
recompilation of "recompile-once" functions. If codegen is enabled, we
now only reset these flags if the execution type is CP and at least the
dimensions are known. This improves the potential of code generation
because the codegen optimizer requires this size information for
checking validity constraints and computing cost estimates.