[SYSTEMML-1835] Improved compilation of sparse-safe codegen cell ops
The codegen templates Cell and MAgg are flagged during compilation as
sparse-safe or -unsafe depending on the characteristics of the specific
operator DAG. This patch makes two minor improvements to the selection
of sparse drivers as well as the analysis of sparse-safe DAGs. 

On GLM binomial-probit (100M x 10, 20/10 outer/inner iterations), this
patch improved the end-to-end runtime w/ codegen from 185s to 145s
(whereas the baseline without fusion requires about ~2,500s).