[SYSTEMML-1941] Fix spark quantile operations w/ non-constant weights
Recently added tests for spark quantile operations (quantile, median,
IQM) revealed issues of incorrect results and other failures. This was
because the respective spark instructions only took the weights into
account for computing the inter-quartile sum but not to find the actual
quantiles. This did not show up before, because most tests only used
constant weights of 1 in order to compare against R, which does not
provide these weighted statistics. 

This patch added the respective tests, and essentially reimplements all
spark quantile operations for quantile, median, and IQM via a new
primitive. The core approach is to compute the sum of weights per
partition, sequentially compute the cumulative sum at the driver to
determine the quantile partitions, and make a final pass to obtain the
actual values as well as partial value proportions. This requires not
just a sorting of values but also a sorted order of partitions and
blocks within each partition. Finally, we now reuse the computation of
corrected IQM values across the CP and spark backends.