[SYSTEMML-1947] Performance cp read sparse binary matrices (seq, par)
A previous performance improvement for reading ultra-sparse binary block
matrices (SYSTEMML-1838) changed the block format of temporary reuse
blocks from MCSR to CSR. For specific scenarios of sparse matrices with
a number of columns lower or equal the blocksize, this caused a fallback
from efficient row copies to the default path of value appends. This was
problematic because it leads to repeated reallocations in the presence
of skew that are unnecessary for matrices with a single column block. 

This patch improves this codepath for all sparse block formats,
including the currently used CSR blocks. By leveraging efficient row
copies without redundant deep copy or reallocations, this modification
significantly reduces GC overhead.

On a scenario of reading the Mnist8m dataset (19GB, sp=0.25) from HDFS,
this patch improved performance from 30.1s (19s GC) to 20.3s (11s GC).