[SYSTEMML-1955] Memory-efficient ultra-sparse matrix multiplication
So far the codepath for ultra-sparse matrix multiply was only used if
any of the two inputs classified as ultra-sparse in terms of its
sparsity and absolute number of non-zeros. For special cases of
permutation matrix multiply with large ultra-sparse matrices this is not
the case leading to sparse-dense matrix multiplication which allocates
the output in dense format. 

This patch improves the code path selection for ultra-sparse matrix
multiplication to include such permutation matrix multiplies, which
allocates the output in sparse format and thus is much more memory
efficient. Additionally, it also improves performance by avoiding
repeated reallocations. 

For example, on a scenario of 58825360 x 2519371 (nnz=58825360) times
2519371 x 300 (nnz=755810998), this patch enables the successful
execution in CP (before it crashed due to allocating a dense output
larger than 16GB). By avoiding reallocations, this patch further
improved the execution time from 112s to 24s due to significantly
reduced GC overhead.