[SYSTEMML-1954] Memory-efficient rexpand output (CSR instead of MCSR)
The rexpand operator is compiled for operations such as table(seq(1,N),
v) and (for the case of seq on the lhs), simply horizontally expands the
given value into the respective column index. For large N, this creates
a huge but ultra-sparse matrix. In past we introduced scalar rows for
the MCSR sparse block to reduce the memory overhead of sparse rows with
a single entry.

However, this approach was still memory-inefficient and causes
unnecessary GC overhead. This patch now creates the output for rexpand
columns in CSR instead of MCSR representation. In order to avoid
unnecessary CSR row pointer updates on append, we also provide the CSR
primitives to construct the block directly from a given column index
array. On a scenario of rexpand 60M rows, this patch improved the
rexpand performance from 39.4s to 3.8s with additional end-to-end
improvements due to reduce GC overhead which also affects subsequent
operations.