[SYSTEMML-1960] Shallow copy on right-indexing of sparse row batches
This patch improves the performance of rightindexing row batches from
sparse matrices with X[rl:ru, ]. In case of our default MCSR sparse
block representation, we now use a shallow copy (pointer move) of rows
because the physical row representation remains unchanged. Result
correctness is ensured because we only use this in the CP matrix
indexing instruction, where other instructions are guaranteed to adhere
to copy on write semantics.