[HOTFIX][SYSTEMML-1959] Fix CSR/COO index search (internal/external pos)
The recent change to use CSR more aggressively for the purpose of
shallow serialize, revealed a hidden issue of CSR/COO index searches,
which are used by various operations that require column-wise
partitioning. For our default MCSR, the internal and external indexes
are equivalent due to separate row objects. This patch fixes the CSR and
COO index searches to return the external indexes (within row), but use
the internal indexes (across rows) for other operations that rely on
these primitives as well.

Furthermore, this includes a number of smaller fixes such as error
handling in grouped aggregate operations, and correct use of sparse
block position offsets in various operations.